Re-instatiate top bar.
Raising the targetSdkVersion means the ActivityBar had to be ported. Done that.
In order to not make things extra convoluted in this new scenario, switching the
content within an activity does not look like a good idea, so I had to refactor a bit,
making this patch rather larger.
