+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/Theme.AppCompat">
+        android:theme="@style/Theme.AppCompat.Light.NoActionBar">
         <activity
             android:name="org.jfet.batsHIIT.HIITMain"
             android:clearTaskOnLaunch="true"
@@ -30,6 +30,16 @@
         <activity
             android:name="org.jfet.batsHIIT.HIITRun"
             android:label="@string/title_activity_hiitrun"
+            android:finishOnTaskLaunch="true"
+            android:screenOrientation="portrait"
+            android:parentActivityName="org.jfet.batsHIIT.HIITMain" >
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="org.jfet.batsHIIT.HIITMain" />
+        </activity>
+        <activity
+            android:name="org.jfet.batsHIIT.HIITRunDone"
+            android:label="Done!"
             android:finishOnTaskLaunch="true"
             android:screenOrientation="portrait"
             android:parentActivityName="org.jfet.batsHIIT.HIITMain" >
