+++ app/src/main/res/layout/activity_hiitmain.xml
@@ -5,7 +5,16 @@
     android:layout_height="match_parent"
     android:id="@+id/main_layout"
     >
-<LinearLayout
+
+	<android.support.v7.widget.Toolbar
+		android:id="@+id/main_toolbar"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"
+		android:background="?attr/colorPrimary"
+		android:minHeight="?attr/actionBarSize"
+		android:theme="?attr/actionBarTheme" />
+
+	<LinearLayout
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="horizontal" >
