+++ app/src/main/res/layout/activity_hiitrun_done.xml
@@ -5,7 +5,16 @@
     android:gravity="center"
     android:id="@+id/hiitRunLayoutDone"
     >
-    <TextView
+
+	<android.support.v7.widget.Toolbar
+		android:id="@+id/rundone_toolbar"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"
+		android:background="?attr/colorPrimary"
+		android:minHeight="?attr/actionBarSize"
+		android:theme="?attr/actionBarTheme" />
+
+	<TextView
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_weight="1"
