--- app/src/main/res/layout/activity_hiitrun_rest.xml
@@ -1,58 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center"
-    android:id="@+id/hiitRunLayoutRest"
-    >
-<TextView 
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:text="@string/rest_string"
-    android:gravity="center"
-    android:textSize="80sp"
-    />
-<TextView
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content" 
-    android:text="@string/zero_string"
-    android:gravity="center"
-    android:textSize="80sp"
-    android:id="@+id/nSecondsRest"
-    />
-<TextView
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:text="@string/second_string"
-    android:gravity="center"
-    android:textSize="24sp"
-    />
-<TextView
-    android:id="@+id/nBlocksRest"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:text="@string/zero_string"
-    android:gravity="center"
-    android:textSize="80sp"    
-    />
-<TextView
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:text="@string/block_string"
-    android:gravity="center"
-    android:textSize="24sp"
-    />
-
-    <Space
-        android:layout_width="match_parent"
-        android:layout_height="54dp" />
-
-    <Button
-        android:id="@+id/pause_button2"
-        style="@style/FolkertButtonText"
-        android:layout_width="300dp"
-        android:layout_height="92dp"
-        android:background="@drawable/btn_green"
-        android:text="@string/pause" />
-
-</LinearLayout>
