+++ app/src/main/res/menu/hiitmain.xml
@@ -1,9 +1,15 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android" >
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android" >
 
     <item
         android:id="@+id/action_help"
         android:orderInCategory="100"
-        android:showAsAction="never"
-        android:title="@string/action_help"/>
+        android:title="@string/action_help"
+        app:showAsAction="never|collapseActionView" />
+    <item
+        android:id="@+id/app_bar_switch"
+        android:title="@string/auto_pause"
+        app:actionLayout="@layout/switch_item"
+        app:showAsAction="never|collapseActionView" />
 
 </menu>
